player search filter so far. recent competitor is not functional.
Due to the unfinished functionality of searching by recent competitor, the player search filter in its entirety is not working as expected (although, taking out all the code for searching by recent competitor will yield correct results, pretty sure :) ).

Excluding the search by recent competitor, an ongoing issues is:
- in order to display any results, both (playerAgeMin and playerAgeMax) and (Country and State) inputs must be filled in. I am not sure why this occurs.

For example, if a only a player name was typed, no results will show as you would need to also fill in the age range and region.